+++ integrations/astroport/packages/abstract-adapter/src/lib.rs
@@ -5,7 +5,7 @@ pub const AVAILABLE_CHAINS: &[&str] = abstract_sdk::core::registry::LOCAL_CHAIN;
 lazy_static::lazy_static! {
     pub static ref AVAILABLE_CHAINS: Vec<&'static str> = {
         let mut v = Vec::new();
-        v.extend_from_slice(abstract_sdk::core::registry::OSMOSIS);
+        v.extend_from_slice(abstract_sdk::core::registry::NEUTRON);
         v.extend_from_slice(abstract_sdk::core::registry::TERRA);
         v
     };