+++ framework/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Ability to set admin to native contracts during instantiation
+
 ### Changed
 
 ### Fixed