+++ framework/CHANGELOG.md
@@ -9,7 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+### Changed
+
+### Fixed
+
+## [0.17.1] - 2023-07-26
+
+### Added
+
 - Ability to set admin to native contracts during instantiation
+- Added neutron
 
 ### Changed